Pull GlslManager and GlslRenderer out into separate files.
