+++ src/commonMain/kotlin/baaahs/glsl/GlslManager.kt
@@ -1,0 +1,7 @@
+package baaahs.glsl
+
+import baaahs.shaders.GlslShader
+
+interface GlslManager {
+    fun createRenderer(program: String, adjustableValues: List<GlslShader.AdjustableValue>): GlslRenderer
+}
