+++ src/jsMain/kotlin/baaahs/glsl/JsGlslManager.kt
@@ -1,0 +1,8 @@
+package baaahs.glsl
+
+import baaahs.shaders.GlslShader
+
+class JsGlslManager : GlslManager {
+    override fun createRenderer(program: String, adjustableValues: List<GlslShader.AdjustableValue>) =
+        JsGlslRenderer(program, adjustableValues)
+}
